+++ agent/go-service/register.go
@@ -12,7 +12,7 @@ import (
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/blueprintimport"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/captureuid"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/attachregex"
-	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/autoaltclick"
+	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/autoalt"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/charactercontroller"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/clearhitcount"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/expressionrecognition"
@@ -56,7 +56,7 @@ func registerAll() {
 	pipelineoverride.Register()
 	expressionrecognition.Register()
 	attachregex.Register()
-	autoaltclick.Register()
+	autoalt.Register()
 	charactercontroller.Register()
 	falseaction.Register()
 	schedule.Register()
